moved all image/file references to the storage driver, added console command to test s3 configurations

+++ app/Services/UserService.php
@@ -15,8 +15,8 @@ use App\Models\User\User;
 use App\Models\User\UserUpdateLog;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Intervention\Image\Facades\Image;
 use Laravel\Fortify\Contracts\TwoFactorAuthenticationProvider;
@@ -289,30 +289,31 @@ class UserService extends Service {
             if (!$avatar) {
                 throw new \Exception('Please upload a file.');
             }
+
             $filename = $user->id.'.'.$avatar->getClientOriginalExtension();
+            $avatarPath = "images/avatars/$filename";
+            $oldAvatarPath = "images/avatars/$user->avatar";
 
-            if ($user->avatar != 'default.jpg') {
-                $file = 'images/avatars/'.$user->avatar;
-                //$destinationPath = 'uploads/' . $id . '/';
-
-                if (File::exists($file)) {
-                    if (!unlink($file)) {
-                        throw new \Exception('Failed to unlink old avatar.');
-                    }
+            if ($user->avatar != 'default.jpg' && Storage::exists($oldAvatarPath)) {
+                if (!Storage::delete($oldAvatarPath)) {
+                    throw new \Exception('Failed to unlink old avatar.'); 
                 }
             }
 
-            // Checks if uploaded file is a GIF
-            if ($avatar->getClientOriginalExtension() == 'gif') {
-                if (!$avatar->move(public_path('images/avatars'), $filename)) {
-                    throw new \Exception('Failed to move file.');
+            $extension = $avatar->getClientOriginalExtension();
+            if (in_array($extension, ['jpg', 'jpeg', 'png'])) {
+                $image = Image::make($avatar)->resize(150, 150);
+                $image->stream();
+                if (!Storage::put($avatarPath, $image)) {
+                    throw new \Exception('Failed to store avatar image.');
                 }
             } else {
-                if (!Image::make($avatar)->resize(150, 150)->save(public_path('images/avatars/'.$filename))) {
-                    throw new \Exception('Failed to process avatar.');
+                $content = file_get_contents($avatar);
+                if (!Storage::put($avatarPath, $content)) {
+                    throw new \Exception('Failed to store GIF avatar image.');
                 }
             }
-
+            
             $user->avatar = $filename;
             $user->save();
